desacople por componentes
