+++ src/app/app.routing.ts
@@ -4,6 +4,8 @@ import { Routes, RouterModule } from '@angular/router';
 // Interal components
 import { HomeComponent } from './components/home/home.component';
 import { AdminComponent } from './components/admin/admin.component';
+import { ListComponent } from './components/list/list.component';
+import { MenuComponent } from './components/menu/menu.component';
 
 const appRoutes: Routes = [
   { path: '', component: HomeComponent },
@@ -16,4 +18,4 @@ const appRoutes: Routes = [
   exports: [RouterModule]
 })
 export class AppRoutingModule {}
-export const RoutingComponents = [HomeComponent, AdminComponent];
+export const RoutingComponents = [HomeComponent, AdminComponent, ListComponent, MenuComponent];
