+++ src/app/class/funciones.ts
@@ -1,0 +1,11 @@
+export class Funtions {
+  constructor() {
+  }
+
+
+  generated_start(number) {
+    const Response = Array.from(new Array(number));
+    return Response;
+  }
+
+}
