+++ src/environments/environment.ts
@@ -4,7 +4,7 @@
 
 export const environment = {
   production: false,
-  API_URL: `http://localhost:3000`,
+  API_URL: `https://angular-hotel-api.herokuapp.com`,
 };
 
 /*
