+++ src/styles.scss
@@ -52,3 +52,10 @@ body{
 .link-color{
   color:#0489B1;
 }
+
+
+.icon-start-category{
+  height: 20px;
+  width: 20px;
+  color: gold;
+}
